Makefile: Create build stamps for images built
For future Travis building, the build and the push steps will be split,
for that reason and because we don't want to rebuild images then a new
build is triggered, stamps are created after an image was built to make
sure they are not build again when pushing.

To clean images there is a new `make clean` target and to make sure to
rebuild images also `make -B` can be used.