Add .travis.yml file
This will build the images as a sort of test and when a tag is created,
the images will be pushed to the DockerHub registry using the
`DOCKER_USER` and `DOCKER_PASS` environment variables for credentials.

The tag of the images is changed from `:vX-DIST` to `:DIST-vX` because
now also for git tags images are tagged with the full git tag:
`:TAG-DIST` and since git tags can contain `-` for pre-releases, it
would have been more complicated to make sense of docker tags.

So, for regular builds the `ORG/IMG:DIST-vX` docker tag is built by
default, and the `ORG/IMG` is tagged for the LATEST_DIST.

When a (git) `TAG` is present, a new docker tag is created with the form
`ORG/IMG:DIST-TAG`. If the `TAG` is a pre-release (contains a `-`), then
the default image is called `ORG/IMG:DIST-vX-test` instead, and no
`ORG/IMG` is tagged.

Fixes #5.