base: Remove the bintray "tools" Debian repo
This repository only hosts tools that are useful for developers but are
not used for building or testing. Having an image with the base
dependencies to build these basic tools that doesn't depend on a bintray
repo is also important to boostrap images for new Ubuntu versions.

This transitively affects the dlang image too.