Create runtimebase image
This image can be used as a base image for running applications (as opposed to
build and test applications, as the current images).

It only installs a very small number of packages and apt repos on top of a base
`ubuntu` image.

In theory the development images could be based on this one, but this
one is kept separated to ease the bootstrapping of new Ubuntu versions,
as to create tools that need to be shipped by the `runtimebase` and
`develdlang` images we need to first have a base development image to
make those packages, so we need to break the circular dependency
somewhere.

Fixes #58.