Use experimental keyserver hkps://keys.openpgp.org
The SKS keyserver network is vulnerable to spam attacks, and these
attacks started to happen. Downloading a spammed key will break GnuPG
installation "in hard to debug ways". To mitigate this problem, switch
to using a new experimental keyserver that is not part of the SKS
network. This server has its own limitations, but it seems to be the way
to go in the future.

For more information about the SKS network attack:
https://gist.github.com/rjhansen/67ab921ffb4084c865b3618d6955275f

Fixes #72.